+++ src/App.jsx
@@ -43,7 +43,7 @@ import OnboardingScreen from "./onboarding/OnboardingScreen/OnboardingScreen.jsx
 import React, { useEffect, useState } from "react";
 import CheckQRCodeScreen from "./onboarding/CheckQRCodeScreen/CheckQRCodeScreen.jsx";
 import { LoginOkScreen } from "./onboarding/LoginOk/LoginOkScreen.jsx";
-import { useAppStart } from "./common/hooks.js";
+import { useAppStart } from "./common/common.hooks.js";
 import { ScanningOptionsScreen } from "./scanning/scanningOptions/ScanningOptionsScreen/ScanningOptionsScreen.jsx";
 import { MainScanningScreen } from "./scanning/scanningSession/MainScanningScreen/MainScanningScreen.jsx";
 import { SourceListScreen } from "./sources/SourceListScreen/SourceListScreen.jsx";
@@ -53,23 +53,23 @@ import {
   personCircleOutline,
 } from "ionicons/icons";
 import { ScanningRecap } from "./scanning/scanningSession/ScanningRecap/ScanningRecap.jsx";
-import { AppContext, UserContext } from "./common/context.js";
+import { AppContext, UserContext } from "./common/common.context.js";
 import { ScanningHome } from "./scanning/scanningOptions/ScanningHome/ScanningHome.jsx";
 import { ScanningNewProfileScreen } from "./scanning/scanningOptions/ScanningNewProfileScreen/ScanningNewProfileScreen.jsx";
 import { ScanningProfiles } from "./scanning/scanningOptions/ScanningProfiles/ScanningProfiles.jsx";
 import { UserProfileScreen } from "./userProfile/UserProfileScreen/UserProfileScreen.jsx";
-import { setDarkModeInDocument } from "./common/util.js";
+import { setDarkModeInDocument } from "./common/common.lib.js";
 
 setupIonicReact();
 
 /**
  * @param {Object} props
- * @param {import("./common/util").DarkMode} props.darkMode
+ * @param {import("./common/common.lib.js").DarkMode} props.darkMode
  * @returns {React.JSX.Element}
  */
 const App = ({ darkMode: initialDarkMode }) => {
   const { data } = useAppStart();
-  /** @type {[import("./common/util").DarkMode, React.Dispatch<import("./common/util").DarkMode>]} */
+  /** @type {[import("./common/common.lib.js").DarkMode, React.Dispatch<import("./common/util").DarkMode>]} */
   const [darkMode, setDarkMode] = useState(initialDarkMode);
   const [userInfo, setUserInfo] = useState({
     instance: { name: "", url: "" },
