+++ src/main.jsx
@@ -3,16 +3,18 @@ import { createRoot } from "react-dom/client";
 import App from "./App";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { IonSpinner } from "@ionic/react";
-import { getPreference } from "./common/preferences.js";
-import { QUERY_KEYS } from "./common/constants.js";
-import { setDarkModeInDocument } from "./common/util.js";
+import {
+  getPreference,
+  QUERY_KEYS,
+  setDarkModeInDocument,
+} from "./common/common.lib.js";
 
 const container = document.getElementById("root");
 // @ts-ignore
 const root = createRoot(container);
 const queryClient = new QueryClient();
 (async () => {
-  /** @type {import("./common/hooks").AppPreferences|null} */
+  /** @type {import("./common/common.hooks").AppPreferences|null} */
   const appPreferences = await getPreference(QUERY_KEYS.APP_PREFERENCES);
   const darkMode = appPreferences?.darkMode ?? "auto";
   setDarkModeInDocument(darkMode);
