+++ src/scanning/scanningSession/CandidateAnnotationsViewer/CandidateAnnotationsViewer.jsx
@@ -3,7 +3,7 @@ import { CandidateAnnotations } from "../CandidateAnnotations/CandidateAnnotatio
 
 /**
  * @param {Object} props
- * @param {import("../../scanningLib.js").Candidate} props.candidate
+ * @param {import("../../scanning.lib.js").Candidate} props.candidate
  * @returns {JSX.Element}
  */
 export const CandidateAnnotationsViewer = ({ candidate }) => {
